passport-photo-ai/backend_backup_working_20251226_103605/requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
rembg==2.0.61
pillow-heif==0.13.0
boto3==1.34.0
python-dotenv==1.0.0
onnxruntime-cpu==1.16.3
mediapipe==0.10.9
hypothesis==6.92.1